\docType{package}
\name{usgs-package}
\alias{usgs}
\alias{usgs-package}
\title{R interface to USGS APIs}
\description{
  R interface to USGS APIs.
}
\author{
  Scott Chamberlain \email{myrmecocystus@gmail.com}

  Karthik Ram \email{karthik.ram@gmail.com}

  Carl Boettiger \email{cboettig@gmail.com}

  Ted Hart \email{edmund.m.hart@gmail.com}
}
\keyword{package}

